Separated the amplitude array constructor into aux
